New Development: No             JIRA Issue: Yes (CAS-4616)         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
FortranizedLoops.h, 
faccumulateToGrid.f, 
AWVisResampler.cc:  Accumulate the CF norm as complex 
                    before using its real part for normalization. 

BeamCalc.cc: A warning removal.

NewMultiTermFT.h: activated getWtImage() method.
NewMultiTermFT.cc: Use GetDoPBCorr() etc. methods instead of the state
                   variables directly.
                   Left some commented tested code.
                   Made initializeToVis() sensitive to doPBCorr().

AWPojectFT.h: virtualized GetDOPBCorr() etc. methods.
AWProjectFT.cc: Removed a few log messages
                if (makingPSF) use cfwts2_p to setup Row2CFMaps.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/code/synthesis@22611 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ TransformMachines/BeamCalc.cc
@@ -580,7 +580,7 @@ namespace casa{
     Int i;
     Double x, freq, df;
     
-    if(0<=ap->band && ap->band<BeamCalcGeometries_p.size()){
+    if((0<=ap->band) && (ap->band<(Int)BeamCalcGeometries_p.size())){
       geom = &(BeamCalcGeometries_p[ap->band]);
     }
     else{
